make more generic usePaginatedqQuery
